Refactor daemon, add autotune to daemon
