+++ lantern_daemon/src/client_embedding_jobs.rs
